[websocket] sets up socket channel for articles list
