+++ client/apps/websocket/messageTypes.js
@@ -1,0 +1,12 @@
+import keyMirror from 'client/lib/keyMirror'
+
+export const messageTypes = keyMirror(
+  'userJoined',
+  'userLeft',
+  'articlesRequested',
+  'userStartedEditing',
+  'userStoppedEditing',
+  'articleAdded',
+  'articleLocked',
+  'articleDeleted'
+)
